+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - iPad support. Status is now displayed at full native resolution on iPad's
 - fixed Sign in: Cannot paste text within password field [#3931]
 
+### Fixed
+- Fixed chat message layout for right-to-left languages
+
 ## [0.9.22] - 2018-07-09
 ### Added
 - Added Farsi public #status channel